Took out control panel functionality out of robot.py and moved it in robotcontainer.py.
the 'controlPanelFunctions' dictionary will hold each function of each button on the control panel. Currently they are unassigned, but will need to be in the future. I have put 21 buttons as of now. This quantity is very likely to change in the future.

The for loop will run 21 times so each button can be initialized as a network button. Currently, there is no prints in the terminal, which a reason for that will need to be figured out.